Zelda functions split across new files. Float references replaced with constants
